requirements.txt
streamlit>=1.22.0
papermill>=2.4.0
mysql-connector-python==8.0.33
ipykernel>=6.25.22
pandas>=1.5.3
matplotlib>=3.7.1
numpy
scikit-learn
SQLAlchemy
openpyxl>=1.24.3
python-dotenv>=1.0.0
streamlit-option-menu>=0.3.2
plotly>=5.15.0
pyecharts
streamlit-echarts